refactor(MapProducerContainer): 지도제작자 컨테이너 컴포넌트 분리
